emit signalReplacePlayerIO() on the client for all non-local players when the server disconnects. the same is already done when a client disconnects from the server. this is necessary to keep a network game working, even if the server disconnects. the clients are then basically in a non-network game and they are server themselves.
furthermore the ID of the players must be fixed: it is not sufficient to just
use the same player ID and replace the KGame ID only. We must use unique numbers
on the client.
(e.g. when there are 2 players in the network game, then both have ID 1, but
with a different KGame ID. when they are now all on the same KGame object, then
both would have the same ID, but that is not valid)

svn path=/trunk/kdegames/libkdegames/; revision=343340