KGameSvgDigits displays a string of digits using SVG themes.
Currently, only the individual digits style works.  I need to do a little bit more
work in code and svg before the 7 & 14 digit themes work.

I'm not that happy with this directory layout, so please do feel free to improve
things if you have more elegant layout. :-)

Note that right now the default size of the digits is HUGE!  This is to help me notice
rendering artifacts.

TODO:
  --7 & 14 digit themes
  --copy ctor, assignment operator
  --better apidox w/ more examples
  --have the test actually test the interface ;-)


svn path=/trunk/KDE/kdegames/libkdegames/; revision=655212