Derive KGameRenderedItem from QGraphicsObject.
Instead of directly deriving from QGraphicsPixmapItem, an internal child
item is used for the rendering. Through clever QGraphicsItem
reimplementations, all input events for the pixmap item are redirected
to the KGameRenderedItem.

This change allows to decouple the coordinate system of pixmap item and
KGameRenderedItem, which will follow in one of the next commits.

svn path=/trunk/KDE/kdegames/libkdegames/; revision=1149327