Remove KExtHighscore classes
They are deprecated, complex, unmaintained and offer lots of functionality we don't use and prefer to not maintain.

The classes were also unfinished porting since kreversi is the only user and there's still no KF5 version of it.

Also they are the last source of kdelibs4support in libkdegames so that's a plus.

The only known user of KExtHighscore is kreversi, see https://git.reviewboard.kde.org/r/124393/ for the review request that adds the files there.

REVIEW: 124392