Fix visual glitch in ace of clubs card in the Paris card deck
Summary:
The ace of clubs card had a bug where the floral wreath surrounding
the ‘clubs’ symbol was instead displayed as basically a big visual mess.
This restores the correct symbol. The symbol used is taken from the
original source of the cards, ‘SVG-cards’, which is available from
https://sourceforge.net/projects/svg-cards/
(I used version 2.0.1, which was released on 2010-07-18.)
Resaving the card deck file with the latest Inkscape version
also greatly reduced its file size, from 3.8 MiB to 1.3 MiB.

BUG: 158390

FIXED-IN: 18.08.2

Test Plan:
Start KPat before and after this commit.
Notice that the design of ace of clubs is now correct.

Reviewers: #kde_games, mlaurent

Reviewed By: mlaurent

Subscribers: kde-games-devel

Differential Revision: https://phabricator.kde.org/D15531