KGamePopupItem: Fix icon and text alignment in HiDPI mode
When running a game in HiDPI mode, the icon within the popup message
would not be vertically centered anymore, and the distance between icon
and text would be too large.

This is caused by the message itself getting too large, due to
calculations using the dimensions of the icon's pixmap in device pixels.

By converting the icon pixmap size to device independent pixels, the
popup message is correctly sized again, leading to correct alignment of
icon and text as well.

Test Plan:

Launch `QT_SCALE_FACTOR=2.4 kblocks`, check centered alignment of icon
and text in popup messages after startup and when pausing/resuming.