With name change, reset library version to follow major Qt/KF versions
Due to the library binary, include dirs and CMake config files having
been renamed to KDEGames6, one could consider this a new library.
The tarballs generated for the sources also contain the KDE Gear version,
not the project one.

So there is no need to continue the old versioning scheme which might be
confusing, due to the offset in major versions.

Also use the KDE Gear version for the patch version, to have some
indication to its release version.