KGamePropertyHandler: use direct connection with QDataStream signal arg
QDataStream (references) cannot be queued, as there is no copy constructor.
Also is it not really defined how a QDataStream instance should handle
calls from different threads.

Avoids the warning by the QObject::connect which otherwise for the auto
connection type also prepares the queued connection and checks for the
registration of the args as meta type.